requirements.txt
asgiref==3.2.3
Django==3.0.3
django-filter==2.2.0
djangorestframework==3.11.0
Markdown==3.2
mysqlclient==1.4.6
Pygments==2.5.2
PyJWT==1.7.1
pytz==2019.3
sqlparse==0.3.0
